Delete items for which there are usage records and no holdings records. Shannon checked and found:
Out of the 2,764 items, 1,677 items are no longer available on the platform.  Of the 1,087 that are still available, I spot checked and it looks like we do not have active holds (I didn’t miss them). As to why we no longer have access, I believe the titles below the trigger point may have been DDA at one time or part of PDA pilot program. On the titles with significant usage, they are not purchased so they must have been subscription, and I assume the publisher pulled them from subscription because they are still available for purchase.  In the last few years, they have been removing titles in June and December with advanced notice and the opportunity to purchase.